refactored to promote SOLID Principles, added first rev pdfParser
